make component loader an inner class
